social_network/requirements.txt
django
djangorestframework
pillow
psycopg2-binary
geopy
djangorestframework-simplejwt